fix: lower coverage threshold to 70% and fix mypy type errors
- Set fail_under = 70 in [tool.coverage.report] (was already 70, confirmed)
- Fix chroma_store.py: add _col property to narrow Any|None type
- Fix chroma_store.py: cast count() return to int to satisfy no-any-return
- Fix sbert.py: remove unused type: ignore on SentenceTransformer import
- Fix sbert.py: handle int|None return from get_sentence_embedding_dimension()
- Fix registry.py: import FAISSVectorStore/ChromaVectorStore from submodules
- Fix registry.py: remove unused type: ignore comments on importlib.util calls

+++ lexisearch/models.py
@@ -207,7 +207,7 @@ class Embedding:
         Returns:
             The Euclidean norm of the vector.
         """
-        return sum(x * x for x in self.vector) ** 0.5
+        return float(sum(x * x for x in self.vector) ** 0.5)
 
     def __repr__(self) -> str:
         """Return a concise string representation."""
